+++ server/src/utils/strings.ts
@@ -1,6 +1,6 @@
 import { FormattingOptions } from 'vscode-languageserver-types';
 
-export function getWordAtText(text: string, offset: number, wordDefinition: RegExp): { start: number, length: number } {
+export function getWordAtText(text: string, offset: number, wordDefinition: RegExp): { start: number; length: number } {
   let lineStart = offset;
   while (lineStart > 0 && !isNewlineCharacter(text.charCodeAt(lineStart - 1))) {
     lineStart--;
@@ -27,25 +27,20 @@ export function removeQuotes(str: string) {
   return str.replace(/["']/g, '');
 }
 
-
 const CR = '\r'.charCodeAt(0);
 const NL = '\n'.charCodeAt(0);
 function isNewlineCharacter(charCode: number) {
   return charCode === CR || charCode === NL;
 }
 
-
-const nonEmptyLineRE = /^(?!$)/mg;
+const nonEmptyLineRE = /^(?!$)/gm;
 /**
  *  wrap text in section tags like <template>, <style>
  *  add leading and trailing newline and optional indentation
  */
-export function wrapSection(text: string, needIndent: boolean, options: FormattingOptions): string {
-  if (needIndent) {
-    const initialIndent = generateIndent(options);
-    text = text.replace(nonEmptyLineRE, initialIndent);
-  }
-  return '\n' + text + '\n';
+export function indentSection(text: string, options: FormattingOptions): string {
+  const initialIndent = generateIndent(options);
+  return text.replace(nonEmptyLineRE, initialIndent);
 }
 
 function generateIndent(options: FormattingOptions) {